requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.0
drf-spectacular>=0.27
django-cors-headers>=4.3
python-dotenv>=1.0
dj-database-url>=2.2
psycopg2-binary>=2.9
whitenoise>=6.6
